* EOAdaptors/PostgreSQLAdaptor/PostgreSQLChannel.m:   fix newValueForDateTypeLengthAttribute * EOAccess/EOEntityPriv.h   declare -_isSingleTableEntity   declare -_assertNoPropagateKeyCycleWithEntities:relationships: * EOAccess/EOEntity.m   fix _flattenAttribute:relationshipPath:currentAttributes:   use relationshipPathBy...Component   make _addAttributesToFetchForRelationshipPath:atts: more understandable   implement -_assertNoPropagateKeyCycleWithEntities:relationships:   implement -_isSingleTableEntity   fix validateValue:forKey: * EOAccess/EOExpressionArray.m   implement -valueWithSQLExpressionElement:forSQLExpression:   fix -valueForSQLExpression: * EOAccess/EODatabaseContext.m   reformat -batchNewPrimaryKeysWithEntity:count:   reformat -prepareForSaveWithCoordinator:editingContext:   reformat and fix -recordChangesInEditingContext:   reformat -recordUpdateForObject:changes:   fix -valuesForKeys:object:   fix -nullifyAttributesInRelationship:sourceObject:destinationObject:   add -_mutableValuesForKeys:object:   add -_recordInsertForIntermediateRowFromSourceObject:...   fix -relayAttributesInRelationship:sourceObject:destinationObject:   fix -relayPrimaryKey:sourceObject:destObject:relationship:   fix -relayPrimaryKey:object:entity:   fix -createAdaptorOperationsForDatabaseOperation:   fix -_buildPrimaryKeyGeneratorListForEditingContext: * EOAccess/EODatabaseOperation.m   clean code * EOAccess/EODatabaseChannel.m   fix -_propertiesToFetch * EOControl/EONSAddOns.[hm]   add NSString (EORelationshipPath)
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gdl2/trunk@37902 72102866-910b-0410-8b05-ffd578937521

+++ EOControl/EOPrivate.h
@@ -160,7 +160,7 @@ static inline void GDL2_AddObjectWithImpPtr(id object,IMP* impPtr,id objectToAdd
 #define GDL2_ObjectAtIndexWithImp(array,methodIMP,index) \
 	(*(methodIMP))((array),@selector(objectAtIndex:),(index))
 
-static inline id GDL2_ObjectAtIndexWithImpPtr(id object,IMP* impPtr,int index)
+static inline id GDL2_ObjectAtIndexWithImpPtr(id object,IMP* impPtr,NSUInteger index)
 {
   if (object)
     {
